// src/interfaces/ICharacter.ts
export interface ICharacterRaw {
  birth_year: string;
  eye_color: string;
  films: string[];
  gender: string;
  hair_color: string;
  height: string;
  homeworld: string;
  mass: string;
  name: string;
  skin_color: string;
}

export interface ICharacter {
  ano_nacimiento: string;
  color_ojo: string;
  peliculas: string[];
  genero: string;
  color_cabello: string;
  altura: string;
  planeta_origen: string;
  peso: string;
  nombre: string;
  color_piel: string;
}