// src/interfaces/IPlanet.ts
export interface IPlanetRaw {
  climate: string;
  created: string;
  diameter: string;
  edited: string;
  films: string[];
  gravity: string;
  name: string;
  orbital_period: string;
  population: string;
  residents: string[];
  rotation_period: string;
  surface_water: string;
  terrain: string;
}

export interface IPlanet {
  clima: string;
  creado: string;
  diametro: string;
  editado: string;
  peliculas: string[];
  gravedad: string;
  nombre: string;
  periodo_orbital: string;
  poblacion: string;
  residentes: string[];
  periodo_rotacion: string;
  agua_superficie: string;
  terreno: string;
}